Move std::os::raw::c_void into libcore and re-export in libstd
